pyproject.toml
[tool.poetry]
name = "freelance-parser"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.20.0.post0"
aiosqlite = "^0.21.0"
alembic = "^1.16.1"
asyncpg = "^0.30.0"
beautifulsoup4 = "^4.13.4"
cloudscraper = "^1.2.71"
requests = "^2.32.4"
python-dotenv = "^1.1.0"
watchdog = "^6.0.0"

[tool.poetry.dev-dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
